feat: extract activity management hook
